Added test for JWT

+++ src/main/java/com/Project/Movie/Movie/MovieNotFoundException.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
 public class MovieNotFoundException extends RuntimeException {
-    public MovieNotFoundException(Long id) {
+    public MovieNotFoundException(int id) {
         super("Movie with ID " + id + " not found");
     }
 }
