+++ src/main/java/com/Project/Movie/Movie/MovieRepository.java
@@ -2,6 +2,6 @@ package com.Project.Movie.Movie;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface MovieRepository extends CrudRepository<MovieEntity, Long> {
+public interface MovieRepository extends CrudRepository<MovieEntity, Integer> {
 
 }
