+++ src/main/java/com/Project/Movie/Movie/MovieServiceImpl.java
@@ -20,7 +20,7 @@ public class MovieServiceImpl implements MovieService{
         return (List<MovieEntity>) movieRepository.findAll();
     }
     @Override
-    public MovieEntity getMovieById(Long id) {
+    public MovieEntity getMovieById(int id) {
         return movieRepository.findById(id).orElseThrow(() -> new MovieNotFoundException(id));
     }
     @Override
@@ -28,7 +28,7 @@ public class MovieServiceImpl implements MovieService{
         return movieRepository.save(movie);
     }
     @Override
-    public MovieEntity updateMovie(Long id, MovieEntity movie) {
+    public MovieEntity updateMovie(int id, MovieEntity movie) {
         MovieEntity existingMovie = movieRepository.findById(id).orElseThrow(() -> new MovieNotFoundException(id));
         existingMovie.setTitle(movie.getTitle());
         existingMovie.setReleaseYear(movie.getReleaseYear());
@@ -37,7 +37,7 @@ public class MovieServiceImpl implements MovieService{
         return movieRepository.save(existingMovie);
     }
 
-    public void deleteMovie(Long id) {
+    public void deleteMovie(int id) {
         MovieEntity movie = movieRepository.findById(id).orElseThrow(() -> new MovieNotFoundException(id));
         movieRepository.deleteById(id);
     }
